requirements.txt
aiogram>=3.4.1
asyncpg>=0.29.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
